requirements.txt
absl-py==2.2.2
accelerate==1.0.0
aiofiles==24.1.0
anyio==4.9.0
argcomplete==3.6.2
async-lru==2.0.5
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.38.14
botocore==1.38.14
brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
dacite==1.7.0
decord==0.6.0
deprecated==1.2.18
diffusers==0.18.2
dill==0.4.0
dnspython==2.7.0
fairscale==0.4.13
fiftyone==1.5.2
fiftyone-brain==0.21.2
fiftyone-db==1.1.7
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
glob2==0.7
graphql-core==3.2.6
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
humanize==4.12.3
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
inflate64==1.0.1
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jsonlines==4.0.0
kaggle==1.7.4.5
kaleido==0.2.1
kiwisolver==1.4.8
lazy-loader==0.4
markupsafe==3.0.2
matplotlib==3.10.3
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==1.39.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
open-clip-torch==2.32.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
patool==4.0.1
pillow==11.2.1
plotly==6.0.1
pprintpp==0.4.0
priority==2.0.0
protobuf==6.30.2
psutil==7.0.0
py7zr==0.22.0
pybcj==1.0.6
pycryptodomex==3.22.0
pydash==8.0.5
pymongo==4.9.2
pyparsing==3.2.3
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
pyzstd==0.17.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rouge-score==0.1.2
rtree==1.4.0
s3transfer==0.12.0
safetensors==0.5.3
scienceplots==2.1.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.46.2
strawberry-graphql==0.268.1
sympy==1.14.0
tabulate==0.9.0
text-unidecode==1.3
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torcheval==0.0.7
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==2.4.0
voxel51-eta==0.14.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
wsproto==1.2.0
xmltodict==0.14.2
zipp==3.21.0
